Split filter dialog into multiple components
Ticket: #247
